alsa: Add udev support
Instead of polling for new devices once per second ask
udev to send us events. Avoids waking up the CPU all the
time and wasting power.

We keep the timeout source as a fallback when udev
support is not enabled, or the udev setup fails for
whatever reason.

The logic for udev_event_ok() was snooped from pulseaudio.